optimisation des filtres
