+++ src/module/statcontainer/index.js
@@ -1,36 +1,15 @@
-import { useCallback, useMemo, useState } from "react";
-import { useRepas, useUser } from "../../utils/hooks/custom";
+import { useCallback, useState } from "react";
 import style from "./style.module.css";
 import RepaCard from "../alimentmacrocard/cardcontainer";
 import TotalMacro from "../totalmacro";
+import { useRepasFilter } from "../../utils/hooks/useRepasFilter";
 
 
 const StatContainer = () => {
     const today = new Date().toISOString().split("T")[0];
     const [dateChoice, setDateChoice] = useState(today);
+    const repas = useRepasFilter(dateChoice);
 
-    const user = useUser();
-    const list = useRepas();
-    const alimentsArray = useMemo(() => {
-        return list.filter((aliment) => aliment.user === user.email);
-    }, [list, user]);
-
-    const repaByDate = useMemo(() => {
-        return alimentsArray.filter((repas) => repas.date === dateChoice);
-    }, [alimentsArray, dateChoice]);
-
-    const petitDèj = useMemo(() => {
-        return repaByDate.filter((repas) => repas.quelRepas === "Petit Dèj");
-    }, [repaByDate]);
-    const déjeuner = useMemo(() => {
-        return repaByDate.filter((repas) => repas.quelRepas === "Déjeuner");
-    }, [repaByDate]);
-    const extra = useMemo(() => {
-        return repaByDate.filter((repas) => repas.quelRepas === "Extra");
-    }, [repaByDate]);
-    const diner = useMemo(() => {
-        return repaByDate.filter((repas) => repas.quelRepas === "Diner");
-    }, [repaByDate]);
 
     const handleChange = useCallback((e) => {
         setDateChoice(e.target.value);
@@ -48,19 +27,19 @@ const StatContainer = () => {
             <ul className={style.statList}>
                 <RepaCard
                     title="Petit Dèj"
-                    repa={petitDèj}
+                    repa={repas.petitDèj}
                 />
                 <RepaCard
                     title="Déjeuner"
-                    repa={déjeuner}
+                    repa={repas.déjeuner}
                 />
                 <RepaCard
                     title="Extra"
-                    repa={extra}
+                    repa={repas.extra}
                 />
                 <RepaCard
                     title="Diner"
-                    repa={diner}
+                    repa={repas.diner}
                 />
             </ul>
         </div>
